src/publishers/telegram_bot/pyproject.toml
[tool.poetry]
name = "telegram_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
confluent-kafka = "^2.3.0"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.7"
aiogram = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
